+++ runtime/src/pallets/mod.rs
@@ -21,3 +21,6 @@ pub use parachain_info::*;
 
 pub mod aura_ext;
 pub use aura_ext::*;
+
+pub mod xcmp_queue;
+pub use xcmp_queue::*;
